requirements.txt
biopython
pandas
numpy
scikit-learn
openpyxl
#later added
requests
matplotlib
keras
tensorflow
keras
freesasa
epitopepredict
torch
tqdm
ruamel.yaml